requirements.txt
coreapi==2.3.3
dj-database-url==0.5.0
Django==2.2.2
django-apscheduler==0.3.0
django-cryptography==0.3
django-filter==2.1.0
djangorestframework==3.9.4
Faker==1.0.7
Markdown==3.1.1
numpy==1.16.4
Pillow==6.0.0
psycopg2-binary==2.8.2
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
tqdm==4.32.2
